+++ Owncloud iOs Client/Login/Login/LoginViewController.m
@@ -1087,7 +1087,7 @@ NSString *loginViewControllerRotate = @"loginViewControllerRotate";
 {
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
     
-    DLog(@"cell tapped number: %d, in section:%d", indexPath.row, indexPath.section);
+    DLog(@"cell tapped number: %ld, in section:%ld", (long)indexPath.row, (long)indexPath.section);
     
     //check if the constant k_hide_url_server is Yes or Not, depend of the branding
     if(k_hide_url_server) {
@@ -1990,7 +1990,7 @@ NSString *loginViewControllerRotate = @"loginViewControllerRotate";
     
     [[AppDelegate sharedOCCommunication] readFolder:_connectString onCommunication:[AppDelegate sharedOCCommunication] successRequest:^(NSHTTPURLResponse *response, NSArray *items, NSString *redirectedServer) {
         
-        DLog(@"Operation response code: %d", response.statusCode);
+        DLog(@"Operation response code: %ld", (long)response.statusCode);
         
         BOOL isSamlServer = NO;
 
@@ -2012,7 +2012,7 @@ NSString *loginViewControllerRotate = @"loginViewControllerRotate";
     } failureRequest:^(NSHTTPURLResponse *response, NSError *error) {
         
         DLog(@"error: %@", error);
-        DLog(@"Operation error: %d", response.statusCode);
+        DLog(@"Operation error: %ld", (long)response.statusCode);
         
         switch (response.statusCode) {
             case kOCErrorServerUnauthorized:
@@ -2063,7 +2063,7 @@ NSString *loginViewControllerRotate = @"loginViewControllerRotate";
  *
  * @warning This method it's present also in AddAcountViewController and EditViewController
  */
--(void)createUserAndDataInTheSystemWithRequest:(NSArray *)items andCode:(int) requestCode {
+-(void)createUserAndDataInTheSystemWithRequest:(NSArray *)items andCode:(NSInteger) requestCode {
       
    // DLog(@"Request Did Fetch Directory Listing And Test Authetification");
     
@@ -2114,7 +2114,7 @@ NSString *loginViewControllerRotate = @"loginViewControllerRotate";
             }
         }
         
-        DLog(@"The directory List have: %d elements", directoryList.count);
+        DLog(@"The directory List have: %ld elements", (unsigned long)directoryList.count);
         
         DLog(@"Directoy list: %@", directoryList);
